Move graph interaction to the utils
