+++ tools/c7n_azure/c7n_azure/resources/key_vault.py
@@ -12,16 +12,16 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from c7n.actions import BaseAction
-from c7n.filters.core import ValueFilter, type_schema
-from c7n.filters.related import RelatedResourceFilter
-
+from azure.graphrbac import GraphRbacManagementClient
 from c7n_azure.provider import resources
-from c7n_azure.resources.arm import ArmResourceManager
 from c7n_azure.session import Session
 
-from azure.graphrbac import GraphRbacManagementClient
-from azure.graphrbac.models import GetObjectsParameters
+from c7n.filters import ValueFilter
+from c7n.utils import type_schema
+from c7n_azure.utils import GraphHelper
+
+from c7n_azure.resources.arm import ArmResourceManager
+
 
 @resources.register('keyvault')
 class KeyVault(ArmResourceManager):
@@ -41,7 +41,6 @@ class WhiteListFilter(ValueFilter):
         if 'access_policies' not in i:
             client = self.manager.get_client()
             vault = client.vaults.get(i['resourceGroup'], i['name'])
-
             # Retrieve access policies for the keyvaults
             access_policies = []
             for policy in vault.properties.access_policies:
@@ -55,10 +54,8 @@ class WhiteListFilter(ValueFilter):
                         'certificates': policy.permissions.certificates
                     }
                 })
-
-            # Enhance access policies with display_name, object_type and principal_name before continue
+            # Enhance access policies with display_name, object_type and principal_name
             i['access_policies'] = self.enhance_policies(access_policies)
-
         return super(WhiteListFilter, self).__call__(i)
 
     def enhance_policies(self, access_policies):
@@ -68,31 +65,14 @@ class WhiteListFilter(ValueFilter):
 
         # Retrieve graph objects for all object_id
         object_ids = [p['object_id'] for p in access_policies]
-        object_params = GetObjectsParameters(
-            include_directory_object_references=True,
-            object_ids=object_ids)
-
-        aad_objects = self.graph_client.objects.get_objects_by_object_ids(object_params)
-        principal_dics = {aad_object.object_id: aad_object for aad_object in aad_objects}
+        # GraphHelper.get_principal_dictionary returns empty AADObject if not found with graph
+        # or if graph is not available.
+        principal_dics = GraphHelper.get_principal_dictionary(self.graph_client, object_ids)
 
         for policy in access_policies:
-            # Ensure there is a graph object. SP can be removed from graph, but it still has an access.
-            if policy['object_id'] not in principal_dics.keys():
-                policy['display_name'] = None
-                policy['object_type'] = None
-                policy['principal_name'] = None
-                continue
             aad_object = principal_dics[policy['object_id']]
             policy['display_name'] = aad_object.display_name
             policy['object_type'] = aad_object.object_type
-            policy['principal_name'] = self.get_principal_name(aad_object)
+            policy['principal_name'] = GraphHelper.get_principal_name(aad_object)
 
         return access_policies
-
-    @staticmethod
-    def get_principal_name(graph_object):
-        if graph_object.user_principal_name:
-            return graph_object.user_principal_name
-        elif graph_object.service_principal_names:
-            return graph_object.service_principal_names[0]
-        return graph_object.display_name or ''
